feat: Introduce MAX_SLIDER_VALUE constant
This commit introduces a constant `MAX_SLIDER_VALUE` in `core.util`. This constant is intended to be used as the maximum value for slider components.

The `EdgeToEdgeSlider.kt` component has been updated to utilize this new constant, although the specific usage within the slider is not shown in this diff.

#29