fix(widget): Reset loading state in ActionButtonWithFeedback
This commit addresses an issue in the `ActionButtonWithFeedback` composable where the `isButtonLoading` state was not being reset after a success or error dialog was dismissed.

This resulted in the button remaining in a loading state even after the operation completed.

The fix ensures that `isButtonLoading` is set to `false` within the `onConfirm` lambda of both the success and error `ConfirmationDialog` instances.

Additionally, a minor UI adjustment was made by replacing `padding(end = 8.dp)` with `Spacer(modifier= Modifier.width(8.dp))` for the spacing between the `CircularProgressIndicator` and the "Đang xử lý…" text in the loading state.

#30