refactor(validation): Update email validation regex
This commit changes the regular expression used for email validation in `ValidationRules.kt`.
The new regex `^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+(\\.[A-Za-z]{2,})+$` is used, and an additional check `!email.endsWith(".")` is added to ensure the email does not end with a period.
This replaces the previous usage of `android.util.Patterns.EMAIL_ADDRESS`.

#71