refactor(home): Replace Scaffold with DrawerWithContent in HomeScreen
This commit replaces the `Scaffold` composable with the custom `DrawerWithContent` composable in the `HomeScreen`.

The `Header` component has been removed from the `topBar` of the `Scaffold` as its functionality is now handled within the `DrawerWithContent` component. The `content` of the `Scaffold` (a `LazyColumn`) is now passed as the `content` parameter to `DrawerWithContent`.

#79