refactor(otp): Remove unused OTP verification and email confirmation states
This commit removes the `verifyOTPState` and `verifyEmailState` `MutableStateFlow`s from `OTPViewModel` as they were not being actively used. The corresponding `verifyOTP` and `confirmEmail` functions have also been commented out.

In `OtpScreen.kt`, the local state variables `sendOTPState` and `verifyOTPState` and the `coroutineScope` have been commented out as they are no longer needed due to the changes in the `OTPViewModel`. The `ConfirmEmailUseCase` injection has also been commented out in the `OTPViewModel`.

The `verifyOTPAndReturnResult` suspend function remains, which directly returns the result of the OTP verification.

#76