feat(auth): Integrate SnackbarViewModel in RecoverPasswordScreen
This commit integrates the `SnackbarViewModel` into the `RecoverPasswordScreen`. The `SnackbarViewModel` is now passed as a parameter to the `HCButton` composable used for the "Gửi" (Send) button.

This allows the button to display snackbar messages, likely for providing feedback to the user after attempting to send a password recovery email.

#85