Refactor(widget): Move Color Options to ColorPicker Composable
This commit refactors the `ColorPicker` composable by moving the `colors` list, which contains pairs of `Color` and `String` (label), directly into the composable function.

Previously, the `colors` list was passed as a parameter to `ColorPicker`. Now, it is defined internally within the function. This change simplifies the usage of the `ColorPicker` by removing the need to pass the color options explicitly from the calling site, as seen in the removal from `ColorPickerPreview`.

The predefined list of colors and their labels now includes:
- Red ("red")
- Green ("green")
- Blue ("blue")
- Yellow ("yellow")
- Cyan ("cyan")
- Magenta ("magenta")
- Gray ("gray")
- Black ("black")
- White ("white")
- Custom Blue (0xFF2196F3) ("customBlue")

#87