Refactor(group): Update `updateGroup` method in `UpdateGroupViewModel`
The `updateGroup` method in `UpdateGroupViewModel` has been refactored to accept an `UpdateGroupRequest` object directly, instead of a `groupName` string.

This change allows for a more flexible and extensible way to update group information in the future, as new fields can be added to the `UpdateGroupRequest` without changing the method signature.

#87